move did_history into did_tdw.core
