+++ did_tdw/core/tests/test_multi_key.py
@@ -1,6 +1,6 @@
 import pytest
 
-from did_history.multi_key import MultiKey
+from did_tdw.core.multi_key import MultiKey
 
 
 def test_multi_key():
